Remove StartupProbe featuregate and related logic
